scalefs: Make the use of mnum_to_name hash-table conditional on debug
We add mnode-to-name mappings on calls to create(), but we don't plan on ever
removing entries from this hash-table (for debugging purposes). So the hash-
table can get pretty heavily loaded, causing performance to collapse.
So remove it out of the regular path by enabling it only in debug mode.