scalefs: Expunge prepare_for/finish_after_commit() and write-lock in transactions
First of all finish_after_commit() doesn't do anything useful since
prepare_for_commit() just asserts that the lock is uncontended; it doesn't
actually return with the lock held. So we can get rid of finish_after_commit()
straightaway.

As for prepare_for_commit(), we always process a given transaction sequentially;
there is absolutely no sharing or concurrent processing involved. So we don't
even need the write-lock to begin with! So get rid of both of them.

And while at it, add an assert statement to the sleeplock code to check that we
are releasing a held lock.