scalefs: Invoke flush_journal_locked() from add_fsync_to_journal()
Today add_fsync_to_journal() uses its own out-of-band path to flush data-blocks
of files to disk. However, there is no need to do that; it can simply add the
transaction to the journal's transaction-log and call flush_journal_locked() on
it. This not only reuses code, but since this will imply using a single path
(via the transaction-log) for all transactions, this will help us a lot with
some upcoming changes that try to parallelize the transaction-logs.