scalefs: Differentiate enqueuing a transaction from writing it to the journal
The journey of an operation logged in the MemFS Oplog looks like this:

Step 1: Apply operation on the buffer-cache blocks to yield a transaction.
Step 2: Add the transaction to the journal data-structure's transaction queue.
Step 3: Dequeue the transaction from the queue and write it to the in-memory
        physical journal along with start- and commit- blocks.
Step 4: Flush the in-memory physical journal (containing the transaction's
        disk blocks) to the on-disk journal, and apply the transaction to the
        on-disk filesystem (DiskFS).

The code doesn't make a good distinction of these steps in terms of
nomenclature. We are going to make a lot of changes to these parts shortly.
So fix some of the nomenclature in preparation.