scalefs: No need to drain the transaction queue in the journal's destructor
The transaction queue is drained once they are committed to the on-disk journal
and applied to the on-disk filesystem. So it doesn't make sense for the journal
class's destructor to try and drain the queue; in fact, it is actually a bug if
the journal's destructor finds the queue to be non-empty!