scalefs: Introduce per-inode-block and per-bitmap-block locks
Today, we hold a global lock (the journal's write_lock) when applying oplog
operations onto the buffer-cache to yield transactions. Going forward, we would
like to remove this sequential bottleneck and apply oplog operations in
parallel.

However, operations that modify the same disk blocks must be ordered carefully:
that is, if op1 and op2 both modify a common disk block b, then if we apply op1
first on the buffer-cache to yield transaction-1, and then apply op2 on the
buffer-cache to yield transaction-2, then these transactions must be also
committed to disk in exactly the same order - transaction-1 followed by
transaction-2, to ensure crash-consistency.

This is trivial to achieve and also automatic when we hold a global lock when
applying oplog operations to yield transactions and then commit them to disk
with that lock still held. But with the global lock gone, we need a different
scheme to preserve this ordering among transactions.

We observe that for any two oplog operations, the only disk-blocks they can
share are: blocks containing the inodes and the blocks containing the bitmap of
used/free blocks, i.e., essentially the filesystem's metadata blocks. Here's
why: The only case where 2 operations can share data-blocks is when both the
operations modify the same file or directory. As for operations that modify a
common directory, we already maintain them in timestamp order using that
directory's oplog, so their ordering is automatically preserved. As for files,
the only operations that touch the file's data blocks are fsync() invocations
on the files, and concurrent fsync()s on the same file are serialized by the
fsync_lock_. So in summary, what this means is that the only disk blocks that
concurrent operations can modify simultaneously are the filesystem's inode
blocks and bitmap blocks.

So, we propose to do fine-grained locking on them - introduce a lock
per-inode-block and per-bitmap-block. Any operation that modifies any of those
shared blocks will acquire the corresponding locks before doing so. That way,
if 2 operations don't modify any common disk blocks, they can be applied to the
buffer-cache truly in parallel to yield transactions! Further, going ahead
they can also be committed to the disk in parallel, since they do not depend on
each other.