scalefs: Add support to acquire and release inode/bitmap locks in a transaction
We plan to do 2-phase locking with the inode-block and bitmap-block locks. That
is, in phase 1, a transaction will try to acquire all the locks of the inode or
bitmap blocks it is going to modify, then perform the modifications, and later
on, in phase 2 (at a suitable point in time), the transaction will release all
the locks it acquired in phase 1. In particular, the number of locks grows
monotonically in phase 1 and shrinks monotonically in phase 2. (That is, as per
standard 2-phase locking, we never release locks in phase 1 and never acquire
locks in phase 2).

Further, we use a simple lock ordering scheme - acquire locks in the increasing
order of the block numbers. This helps us avoid deadlocks as well as retries
of the 2 phases of locking. The reason why we have the luxury of doing this is
rather subtle, as explained below.

For every MemFS operation, we know beforehand the set of inodes that it can
potentially modify; so we can sort and acquire the inode-block locks before
performing the modifications. As for bitmap-blocks, we won't know beforehand
which blocks we will modify. But luckily for other reasons, when building the
transaction, we don't update the buffer-cache's bitmap blocks immediately to
note down the blocks allocated or freed by the transaction; instead we store
the updates to the bitmap in the transaction's allocated_block_list and
free_block_list, and after all the modifications are done, these lists are used
to update the buffer-cache's bitmap blocks. What this means is that the shared
bitmap blocks in the buffer-cache are updated only after we know exactly which
bits are to be flipped on or off! And that's exactly what we want - to know
beforehand which bitmap blocks will be modified, before actually performing
the modifications, so that we can acquire the corresponding locks. This is why
it all works out well, and we never have to retry 2-phase locking :-)

So a transaction's sequence of actions will look something like this:

  - We know which inodes might be modified by looking at the operation itself.
  - Acquire the inode-block locks corresponding to the inodes.
  - Modify inode blocks and data blocks in the buffer-cache, building lists of
    blocks allocated and freed by this transaction.
  - (As of this point, the inode blocks are the only shared blocks that have
    been modified, with their corresponding inode-block locks held).

  - We now know which bitmap blocks will have to be modified, by looking at the
    lists we built above.
  - Acquire the bitmap-block locks corresponding to these block numbers.
  - Modify the bitmap-blocks in the buffer-cache with their corresponding locks
    held.

So overall, every shared block is always updated with its lock held.

An important question here though, is when to perform phase 2: i.e., when is it
safe to release the locks? So far, we have a single transaction queue (as we have
a single journal), so the moment a transaction is queued to the global transaction
queue, its order with respect to other transactions/operations is set in stone
and will be preserved throughout, all the way until it commits to the disk!
So we can drop the acquired locks right after we successfully enqueue a
transaction to this global queue. So that's where we perform phase 2.

Add the necessary support to implement this feature, so that every transaction
can make use of it. For convenience, we expose APIs which accept a list of inode
numbers or data-block numbers; and the API calculates their corresponding inode-
block and bitmap-block numbers under the covers, deduplicates them (so that we
don't double-acquire the lock of a single block) and then acquires the appropriate
set of unique inode-block or bitmap-block locks.