scalefs: Acquire bitmap-block locks in pre_process_transaction()
pre_process_transaction() updates the bitmap blocks in the buffer-cache. So
teach it to acquire all the corresponding bitmap-block locks before modifying
the buffer-cache.

Also, the current code invokes pre_process_transaction() after the transaction
is enqueued in the global transaction queue. That is too late for us; we need to
acquire the bitmap-block locks before enqueuing the transaction, so arrange
pre_process_transaction() to be invoked earlier.