scalefs: Apply oplog operations on the bufcache in parallel to yield transactions
We are all set to apply oplog operations on the bufcache without holding the
journal's write-lock. Fine-grained locking on the inode-block and bitmap-block
locks will provide safety along with scalability for these transactions. So
do the final step of moving the journal's write-lock acquisition to the point
after applying the oplog operations :-)