fs: Increase the filesystem size to 16 GB
Today our filesystem size is 4 GB, which gives us about 32 bitmap blocks. We
would like to allocate bitmap blocks per-cpu, so increase the filesystem size so
that we can have more of them. With a 16 GB filesystem, we'll get around 128
bitmap blocks, which is more than enough for our current target of 80 CPUs.