fs: Add infrastructure to support a simple per-CPU inode (inum) allocator
This is based on the design of the recently added per-CPU block allocator.
Specifically, we divide the available free inode numbers among the CPUs and
initialize their respective freelists at boot time. With this new allocator,
tasks on different CPUs can allocate and free inodes truly in parallel. As long
as each CPU has enough free inode numbers of its own, there won't be any cache-
line sharing or contention across CPUs while allocating and freeing inodes.

Also, we add a global reserve pool of inode numbers, using whatever is left
over after evenly dividing the free inodes among all the CPUs. This would be
the first fallback strategy for inode-number allocations if a per-CPU freelist
becomes empty.

This is both a scalable and a high-performance allocator; all allocations and
frees can be done in constant time.

[ Where the distinction between inodes and inode numbers matters, strictly
speaking, this is an inode-number allocator. It doesn't allocate inodes, the
in-memory objects encompassing all the fields of an inode. ]

This patch adds the basic infrastructure and the initialization code; subsequent
patches will add the remaining bits.