fs: Free inode numbers only upon transaction commit
Just like disk blocks should not be reused for other transactions until the
transaction that frees the blocks is successfully committed to the disk, we
should also take care not to reuse inodes prematurely. So add the necessary
infrastructure to free inode numbers in the in-memory bitmap after the
transaction freeing that inode commits successfully.