scalefs: Fix false positives that lead to deferring inode reclaim
When the last link of a file or directory is removed, and there are no open file
descriptors referring to it, then we should be able to free the inode
immediately rather than deferring the reclaim until reboot. In the current code,
there are a bunch of factors that trick us into going down the deferred reclaim
path even when it is safe to immediately free the inode and its resources.

Now this might sound somewhat dull, but it has some rather important consequences.
If an inode isn't freed immediately when its last unlink is flushed to the disk,
both its disk blocks and its inode number cannot be reused until reboot. This
means that the per-CPU block and inode allocators will prematurely exhaust their
per-CPU freelists due to the leak, and try to tap into the global reserve pools;
which in turn causes heavy contention, thus hurting scalability drastically.
This was observed in the parallel LFS Smallfile benchmark runs.

Avoid this catastrophe by fixing the underlying issues. First,
process_metadata_log_and_flush() creates a list of references to dirty mnodes
before calling process_metadata_log(). Those references bump up the refcounts,
which were not being accounted for when checking if the inodes can be freed
immediately. So fix it by maintaining a list of mnode numbers instead. Secondly,
the output of get_consistent() on a file or directory right after unlink is
actually 2, not 1 (most likely due to the additional refcount used to pin it in
the mnode cache, which will be unpinned when the linkcount eventually hits its
onzero() method), so fix that check as well. Also, unlinked directory mnodes were
not being marked as clean even after their oplogs were flushed, so freeing their
entries from the metadata_log hash-table used to cause a crash in
process_metadata_log() because it thought that those directories were dirty and
was trying to flush their (non-existent) oplogs again! Fix that as well.