scalefs: Prepare to add per-CPU journals by passing the CPU as argument to functions
Soon, we are going to add per-CPU transaction-queues and per-CPU journals, for
scalability. So during sync/fsync, we'll want to make sure that we add the
transactions to the local CPU's transaction-queue and then flush that CPU's
journal to disk.

However, since threads can get migrated midway, if we are not being careful, we
might end up adding the transaction to CPU x's transaction-queue and then flush
out CPU y's journal (due to thread migration). To avoid that, ascertain the CPU
we are running on right at the beginning of the sync/fsync and pass it along as
a parameter to the entire code flow that follows, so that it can proceed
undisturbed by process/thread migration.