scalefs: Restructure the locking around enqueue_transaction() a bit
Going ahead, we would like to have a critical section where we hold all the
locks we need in order to set up data-structures to track dependencies between
transactions that get applied to disk via different journals. So restructure
the locking code a bit to prepare the ground for this critical section.

Specifically, we want to hold all the inode-block and bitmap-block locks
(Phase 1 of the 2-Phase locking) and also the lock guarding the per-cpu
journal's transaction queue.