scalefs: Note down the timestamp and queue id where each transaction is queued
This will help us in tracking dependencies between transactions that modify
common disk blocks, so that even if they are committed and applied to disk via
different journals, we can ensure that they are applied in the correct order.

Note that we need the timestamps of the transactions going into a given queue
to be monotonically increasing (so that we can wait for dependent transactions
to be applied by looking at which timestamp has been processed through). That's
why we introduce a new field called 'enq_tsc', instead of reusing the creation
timestamp of the transaction, because the latter cannot guarantee this property.

While at it, also maintain a list of inode-blocks and bitmap-blocks that each
transaction modifies. This will be useful in upcoming patches.