scalefs: Add a hash-table to track dependent transactions going into different queues
If two transactions t1 and t2 modify the same disk blocks, then if t1 updated
the blocks before t2, t1 must also be applied to the on-disk filesystem before
t2. However, t1 and t2 might get enqueued to different journals if they happened
to be executing on different cpus at the time of enqueuing.

So, to enable t2 to wait until t1 is applied to disk, at a minimum, t2 would
need to know which queue t1 entered. This patch adds a hash-table to enable
this tracking. Specifically, given any common disk block that transactions can
modify (i.e., either an inode block or a bitmap block), the hash-table keeps
track of the last transaction that modified that block, by way of maintaining
its queue id and the timestamp at which it was enqueued to that particular
transaction queue.