scalefs: Add infrastructure for each transaction to keep track of its dependencies
If transaction A depends on transaction B, then after committing them to their
respective on-disk journals (perhaps in parallel), we need to be able to postpone
applying transaction A to the on-disk filesystem until transaction B is applied.
So add a vector to each transaction to keep track of which journal queue (and
transaction timestamp) it depends on.