scalefs: Control the transaction batching window to avoid deadlocks
We flush transactions in batches to the on-disk journals to improve performance.
However, if transactions that depend on each other are committed via different
journals, then we need to carefully control which transactions go in the same
batch, in order to avoid deadlocks. The following example illustrates the problem:

Consider the following scenario, where we have 3 transaction queues
corresponding to 3 on-disk journals. Transactions 101, 102 and 103 will be
flushed via Queue 1 and so on.

    Txn Queue 1    Txn Queue 2    Txn Queue 3

 |   txn 103        txn 203        txn 303
 |   txn 102  <---  txn 202        txn 302
 |   txn 101  --->  txn 201        txn 301
 V

Suppose transactions 101, 102 and 103 are being prepared to be flushed in a
single batch, and similarly assume that 201, 202 and 203 are also being prepared
to be flushed in a single (separate) batch. Now, if txn 101 depends on txn 201
and txn 202 depends on txn 102, we would have a circular dependency, and hence a
deadlock! In other words, there would be simply no way to flush these 2 batches
to the disk! Of course, in practice, these cross-queue cyclic dependencies could
be much more complicated, extending to multiple journal queues.

However, observe that, given any 'n' individual transactions that depend on each
other, there can be no cycle in the dependency chain, because they can be totally
ordered by their enqueue-timestamps (timestamps at which they were enqueued into
their respective journal queues). The problem above is that batching too many
transactions together can lead to cycles in the dependency chain, since each batch
is to be flushed as a single unit, and there is no guarantee that we can find a
total ordering between these batches, as opposed to individual transactions.

We propose to avoid these deadlocks by closing the batching window before the
number of cross-queue dependencies exceeds by 1. We achieve this by enforcing a
rule, that for any batch of transactions, only the first transaction in the batch
is allowed to have a cross-queue dependency. All other transactions in the batch
can only depend on their predecessors and no one else. In terms of cross-queue
dependencies, this is logically equivalent to having a single transaction being
flushed from each queue at a time, which we showed to be deadlock-free above
(because of their total ordering).

Thus, this rule helps us maintain the invariant that any batch of transactions
depends on at most one other batch of transactions in a different queue. Further,
these dependencies are guaranteed to be in 1 direction only (no cycles), since
only the head transaction in each batch is allowed to have cross-queue
dependencies.

So, in the above example, transactions 101, 102 and 103 can all be flushed in
a single batch since only 101 depends on another queue. However, in queue 2,
transaction 201 will be flushed individually since 202 adds a dependency on
queue 1 and only head-of-the-batch transactions are allowed to have such
dependencies. And after 201 is flushed, 202 and 203 can go together in a single
batch. This selection of transaction batches avoids the deadlock between queues
1 and 2.

Also, with this change, the timestamp of a transaction batch (as written in
the 'start' and 'commit' records in the journal) is the same as the enqueue-
timestamp of the first transaction in that batch. [ This will be useful for
both applying that batch of transactions and recovering them (in case of a
crash) in later patches ].