scalefs: Postpone applying a transaction until its dependencies are met
Dependent transactions can be safely committed in parallel to multiple on-disk
journals. However, we need to honor their dependencies when applying them to the
on-disk filesystem. So add the infrastructure to postpone applying a transaction
batch until all its cross-queue (cross-journal) dependencies are met.

As far as cross-queue dependencies are concerned, we need to only look at the
first (head) transaction in each batch, since that's the only transaction
allowed to have any cross-queue dependencies. On the other hand, we must call
notify() using the latest timestamp in the batch, namely the timestamp of the
last transaction, so that all the waiters can be woken up in a timely manner.