scalefs, jrnl recovery: Apply dependent transactions in the right order after crash
After a crash, each on-disk journal can have at most 1 committed transaction
that wasn't applied to the on-disk filesystem. Dependencies between these
transactions (if any) can be easily resolved by ordering them by their encoded
timestamps. So to keep things simple, we sort all these committed transactions
by their timestamps and conservatively apply them in order after reboot,
irrespective of whether they were actually dependent on each other.

Note that the timestamps encoded in the commit record in each of the journals
is the enqueue-timestamp (enq_tsc) of the head (first) transaction of the batch
that was last committed to that journal (as opposed to the timestamp of the last
transaction of that batch). This helps us ensure that sorting them by their
timestamps preserves their dependency order, since only the head transactions
are allowed to have cross-queue (cross-journal) dependencies.