fs: Initialize per-CPU inum allocator uniformly after offsetting allocated inums
The first few inode blocks don't contain any free inums because we would have
allocated them to the initial set of files and directories in the filesystem.
So, if we divide up the inums equally among the CPUs starting from inum 1, the
first few CPUs can end up getting no free inums at all, if most of them have been
used already. (So, right from the beginning, the first set of CPUs might start
allocating from the global reserve pool, hurting scalability). This has not been
a problem yet for the per-CPU inum allocator, but seeing how this same
underlying issue caused problems in the per-CPU block allocator, it would be
better to fix the per-CPU inum allocator as well, while we are at it.

Fix this by initializing the per-CPU inum allocator by offsetting the first few
inode blocks that have no free inums, so that CPU0 and friends don't start with
no resources of their own.