scalefs: Postpone committing a transaction until its dependencies are met
We already take care to postpone applying a committed transaction (batch) to
the on-disk filesystem until all the transactions it depends on (perhaps
committed via different journals) are applied to the disk filesystem.

It turns out that we must also use a similar postponement logic when committing
transactions too! For example, suppose transaction A depends on transaction B,
but they are in different journal queues. Then, if transaction A commits before
B and then the system crashes, the recovery code (upon reboot) will end up
applying A, and forgetting all about B (because it didn't get a chance to commit
to its journal), which would be incorrect. So fix this by honoring dependencies
between transactions when committing them, just like we do when applying
committed transactions to the on-disk filesystem.