scalefs: Absorb file link/unlink MemFS operations that cancel each other
If a directory's mfs-log has a link and an unlink of the same link-name, they
effectively cancel each other, so we don't have to flush them out to disk.
Implement this absorption of link and unlink operations for files. Absorbing
directory operations is more complicated, so we'll postpone it for now.

Also, our fsync algorithm for handling renames (especially cross-directory ones)
is rather complex and fragile, and it is very likely to break (not only the
algorithm but also our invariants) if we cancel links and unlinks that cross
a rename boundary. So be conservative and absorb operation pairs that match
only if they are not separated by renames. Also, retry the absorption after a
rename is successfully processed, so that we don't miss out on valid and safe
opportunities for doing absorption.