scalefs: Get rid of the 'count' variable from the loop in process_ops_from_oplog()
Processing the 'create' operation of an mfs_log (the very first operation, if it
is yet to be flushed) can be handled much more cleanly than having an awkward
count variable which has almost no impact on the loop itself. So rewrite this
part by moving the processing of 'create' operations outside of the loop.