scalefs: Remove the reclaim-inodes list from the superblock in prep for a rewrite
Keeping a list of potentially unreachable inodes (to be deleted/reclaimed on
reboot) in the superblock was a stop-gap solution with several limitations: it
could only keep a very small set of inodes (about 1020), and was also a
sequential bottleneck.

Remove the whole thing, in preparation for introducing a much more scalable
infrastructure that also doesn't enforce limitations on the size of the list.