scalefs: Add infrastructure to mark potentially unreachable inodes for deletion
A newly created file, even if fsynced, remains unreachable on the disk until its
link in the parent is flushed by fsyncing the parent directory. So, if the
system crashes before the parent is fsynced, the file must be deleted during
crash-recovery on reboot.

To enable this, reserve a dedicated "inode-reclaim" file on the filesystem (just
like the on-disk journal) to keep a list of potentially unreachable inodes (that
are to be reclaimed on reboot).

To keep things efficient when adding or removing entries from this files, we use
the same technique that we use for managing on-disk directories - use an in-memory
hash-table that maps inode numbers to their offsets in the file. This helps us
avoid linear searches when reviving inodes by removing them from this list.
To add inodes to the list, we simply append them to the file, without reusing
any empty slots.

In principle, multiple transactions can update this inode-reclaim file
concurrently, which means that we would have to do two-phase locking for this
file, in order to synchronize the updates and also ensure that the ordering
between transactions that update this file is preserved even when committing and
applying those transactions to disk.

However, we observe that we already do two-phase locking for inode-blocks and
bitmap blocks. In particular, for every MemFS operation, we know the set of
inodes that it might potentially modify on the disk, and we acquire inode-block
locks corresponding to those inodes, and we manage those locks using the
two-phase semantics. Luckily, it is also true that the set of inodes that an
operation might potentially want to mark as unreachable, is always a subset of
all the inodes that the operation might modify. Since we already perform
two-phase locking for those inodes (the superset), we don't need any additional
locking to protect updates to the inode-reclaim file.