scalefs: Reclaim unreachable files and (empty) directories during crash-recovery
During crash-recovery, delete the inodes (and their metadata and data blocks)
which have been marked as unreachable. We handle the easier case of deleting
files and empty directories for now. Reclaiming unreachable non-empty directories
is more complicated, because we'll have to recursively unlink/delete their files
and child directories too. So we'll deal with that later on.