scalefs: Add per-CPU inode-reclaim files for scalability
Having a single on-disk inode-reclaim file is obviously a bottleneck for
multi-core scalability. So add per-CPU copies of that file.

This introduces some rather unexpected issues which must be handled carefully.
For example, a newly created file might be fsynced on CPU x and then its parent
directory fsynced on CPU y. In this case, the new file's inode will be added to
CPU x's inode-reclaim file during the first fsync, and then during the second
fsync it must be revived (removed) from the same inode-reclaim file even though
the second fsync runs on CPU y!

To support this, add a hash-table to maintain a mapping of dead-inodes to the
CPUs whose inode-reclaim files they have been marked in, so that we can remove
an inode from its file if we want to revive the inode at a later point in time
(because it is now reachable again, say after fsyncing its parent). The
hash-table lookup would help us perform the revival correctly, irrespective
of the CPU on which it is run.

[ In principle, a simpler approach would have been to have an add-list and a
remove-list for every CPU, both of which grow monotonically, so that every CPU
can mark or revive inodes using its local inode-reclaim file alone. However, in
practice, benchmarks like LFS Smallfile can make these lists grow unbounded,
making it very hard to run a large number of iterations of the benchmark without
exhausting the fixed size of these on-disk files. So we use the approach outlined
above instead. ]