scalefs: Mark newly created inodes as unreachable during fsync
If a newly created inode (file or directory) is fsynced, it remains unreachable
on the disk at least until its link in the parent is also flushed by fsyncing
its parent. So mark such inodes as potentially unreachable and hence eligible
for deletion/reclaim upon reboot.

On the other hand, when a link to the file or directory is being flushed to
disk, revive the inode by removing it from the inode-reclaim list, because it is
now reachable on the disk (assuming that its parent directory is reachable
itself, of course).

It turns out that we can actually revive the inode unconditionally upon flushing
a link to it, without really checking if its parent directory is reachable on
the disk, because if the parent was unreachable, it would have been added to the
inode-reclaim list when it was fsynced. And the crash-recovery code is
responsible for recursively reclaiming all the files and sub-directories of an
unreachable but non-empty directory which had been added to the inode-reclaim
list; so it all works out pretty well.