scalefs: Don't modify the 'size' variable of the inode-reclaim files
The inode-reclaim files are supposed to be of fixed size, so we shouldn't really
update the 'size' variable in their inodes when adding or deleting inode numbers
to/from the list.

Further, when reclaim_unreachable_inodes() is called during boot, even if there
is no inode to be reclaimed, it wastes time flushing out a transaction that
resets the size to 0 for every per-CPU inode-reclaim file. Fix this by never
modifying 'size' in the first place. The 'next_offset' variable is sufficient to
manage addition/removal of inodes in these files.

And while at it, also fix reclaim_unreachable_inodes() so that it removes the
inodes which were reclaimed, so that we don't try to redo them on the next boot.