scalefs: Deduplicate blocks before checking if the transaction fits in journal
We perform deduplication in conjunction with group commit, after building a set
of transactions that we want to flush together, so that we can reduce the number
of disk writes if many of those transactions update the same blocks. But this
deduplication happens after checking if each individual transaction fits in the
journal by itself.

However, for usecases like the LFS Largefile benchmark, where we accumulate a
lot of (repeated) metadata blocks in a single transaction, it might be hard or
even impossible to fit the transaction in the journal if we don't deduplicate
it first! So add a deduplication round for individual transactions too.