diskfs: Skip synchronous disk reads in bmap() whenever possible
bmap() allocates blocks if the inode doesn't have a block already at the
specified offset. So when it allocates an indirect block or a first-level or
even a second-level doubly-indirect block, we don't have to read the block from
the disk, as it was freshly allocated and doesn't have anything on it yet.

So find all such opportunities in bmap() and skip synchronous disk reads
whenever possible. (But make sure to reset things to default appropriately, so
that we don't skip disk reads on the wrong blocks!).

This helps improve the performance of LFS Largefile to some extent.