scalefs: Don't delete inode if link count drops to zero in the rename path
A rename is performed atomically, and any changes to the link count of the inode
during rename is only temporary and is not visible after the rename commits (or
aborts). So we shouldn't delete the inode on the disk if the link count drops to
zero during renames.

This isn't actually a bug at the moment, because we process mfs-rename-link
first, before processing mfs-rename-unlink. But this patch provides a cleaner
way to handle renames.