scalefs: Fix long standing bug in processing link/unlink operations from oplog
Consider the following sequence of operations:

In MemFS:

 - Create file 'f' in dirA.
 - Link 'f' into dirB.
 - Unlink 'f' from dir A.

Then, fsync file 'f'.

Due to the fsync, a new inode is allocated to file 'f' and the file contents are
persisted on the disk. Next, if dirA if fsynced, the current code deletes file
'f' on the disk, because it was linked and then unlinked from dirA (and hence
the link count of file 'f' as calculated locally is 0). But this is clearly
wrong, because there is still an outstanding link yet to be flushed from dirB!

This problem occurs because, when fsyncing a particular directory, it knows only
about local information as seen from that directory's oplog; that is, the code
doesn't have any knowledge of how many links to the file there are in total,
including those which might be waiting to be flushed in other directory oplogs.
And thus, deciding whether to delete the file on disk based purely on this local
link count calculation can lead to incorrect results. To fix this, what we need
is a hand-shake mechanism between MemFS and fsync about mnode link-counts, so
that we can accurately determine when it is safe to delete inodes on the disk.
This solution is outlined below.

We introduce a per-mnode link-counter to keep track of the global link count of
that mnode, and both MemFS and fsync/sync code are involved in maintaining the
link-count via a hand-shake, as depicted below.


         MemFS                   |         Fsync/Sync
                                 |
 Upon adding a new link to       |  Upon processing an unlink on the inode,
 mnode m, increment m's          |  decrement the mnode's link-count. We do
 link-count. It is never         |  this in both cases: when an unlink is
 decremented on any operation    |  flushed to disk and when a link+unlink pair
 in MemFS, including unlinks.    |  is absorbed in the oplog.
                                 |

In this scheme, whenever the link-count drops to zero in the fsync/sync path, it
really does indicate that there no more outstanding links to that inode waiting
to be flushed from any of the per-mnode oplogs; and hence it is safe to delete
the inode and its file-contents from the disk, and this solves the problem
described above. [ There are a few corner cases though, that we need to handle
carefully here: for example, if the system crashes right after fsyncing dirA,
but before fsyncing dirB, the on-disk link-count of file 'f' would be zero, and
hence it should be reclaimed on reboot. We take care of this by using the
inode-reclaim infrastructure that we have already built. ]

This per-mnode link-counter should not vanish when the mnode goes away; it must
be available to the fsync code until all the directories into which that mnode
was linked are flushed. To that end, we store this per-mnode link-counter
alongside that mnode's oplog, and thereby decouple its lifetime from that of the
mnode data-structure.

If we just had a single global counter per-mnode, concurrent links to that file in
MemFS wouldn't scale. So we use a per-CPU, per-mnode counter instead. However,
for accurate zero-detection, we add up all the per-CPU counts whenever we
process unlinks in the fsync/sync path. This doesn't worsen fsync scalability
beyond what it already is, because we can't process concurrent unlinks to the
same file in the fsync path anyway, because all of them have to call iupdate()
on the same file inode, which is already a sequential bottleneck. Thus, this
per-CPU link-count infrastructure merely fixes the bug, without hurting
scalabilty in any part of the system.

Finally, a small note on the implementation of the counter: we develop a simple
per-CPU counter from scratch instead of using refcache, because in my
experiments, refcache's get_consistent() doesn't seem to be very reliable; it
seems to give rather inconsistent results.