scalefs: Absorb 'create' or delete inode when absorbing its last link and unlink
When absorbing link and unlink operations, if the global link-count of the inode
drops to zero, and the inode is already on the disk, make arrangements to delete
the inode. Otherwise, if the mnode's 'create' operation hasn't been flushed yet,
absorb it just like we absorbed the link and unlink operations.

[ Note: In the former case mentioned above, one could imagine skipping the delete
and postponing it until reboot, since we are only absorbing link-unlink pairs
and not flushing them to disk. However, if we don't delete the inode and free up
its resources right-away when the opportunity arises, we'll start depleting
usable inode numbers and associated resources, making benchmark iterations run
progressively slower ].

This is easier said than done because weaving this logic into the existing
framework of how we process the metadata logs is rather non-trivial. For
instance, we should be careful not to acquire mfs-log locks or sync_locks of
multiple oplogs at once, because there would be no way to get that lock ordering
right, which might lead to deadlocks. Hence, it is impossible to check whether
the 'create' operation of an inode has been flushed or not while we are
absorbing the link/unlink of that inode. Further, even if we somehow knew that
the inode had already been persisted on disk, it is too early to really delete
it at the time of absorption, because we wouldn't have a transaction yet, and we
wouldn't be able to safely release the mfs-log and mnode-lock resources of that
inode without corrupting memory, because the rest of the loop may access them.
Instead, we solve/avoid all of these issues in one go by postponing the
absorption of the 'create' or the deletion of the inode from the disk, until all
the other oplog entries are processed in the current invocation of
process_metadata_log().